+++ component-library/src/data-contracts/financial-service/data-contracts.ts
@@ -3566,15 +3566,15 @@ export type TriggerSyncData = any;
 
 export interface TriggerSyncRequest {
   /**
-   * the account id associated with the user
-   * @format uint64
-   */
-  financialAccountId?: string;
-  /**
    * tha financial account type
    * @example "username:testuser"
    */
   financialAccountType: FinancialAccountType;
+  /**
+   * the account id associated with the user
+   * @format uint64
+   */
+  linkId?: string;
   /** @example "username:testuser" */
   profileType: FinancialUserProfileType;
   /**